# internal/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: lowcode-database API
  description: |
    Postgres-backed low-code table service: dynamic tables, columns, rows, indexes,
    relations, and saved queries.

    Optional HTTP webhooks: `/v1/admin/webhooks`. Out of scope (schema bundle import,
    plugins, graph expand, ENUM, RBAC): see docs/roadmap.md.

    **Multi-tenant:** send `X-Tenant-Id` on every request when `TENANT_MODE=multi`.
    In single-tenant mode the header is optional (default tenant from env).

    **Playground:** [lowcode-database-playground](https://github.com/monoposer/lowcode-database-playground) — AG Grid UI (separate repo).

    Errors are returned as plain text (`http.Error`), not JSON envelopes.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local dev (see HTTP_ADDR in .env)
...
paths:
...
  /v1/data/tables/{tableName}/rows:
    get:
      tags: [Tables]
      summary: List rows (paginated)
      operationId: listRows
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
          description: Also accepted as `page_size`.
        - name: pageToken
          in: query
          schema:
            type: string
          description: Also accepted as `page_token`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListRowsResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    post:
      tags: [Tables]
      summary: Create row
      operationId: createRow
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRowBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRowResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/data/tables/{tableName}/rows/{rowId}:
    patch:
      tags: [Tables]
      summary: Update row
      operationId: updateRow
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
        - $ref: "#/components/parameters/RowIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRowBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateRowResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Tables]
      summary: Delete row
      operationId: deleteRow
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
        - $ref: "#/components/parameters/RowIdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteRowResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/data/tables/{tableName}/rows:query:
    post:
      tags: [Tables]
      summary: Query rows with filter, sort, and projection
      operationId: queryRows
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRowsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRowsResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/data/tables/{tableName}/rows:bulkUpsert:
    post:
      tags: [Tables]
      summary: Bulk upsert rows
      operationId: bulkUpsertRows
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpsertRowsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpsertRowsResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/data/tables/{tableName}/rows:bulkDelete:
    post:
      tags: [Tables]
      summary: Bulk delete rows
      operationId: bulkDeleteRows
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/TableNamePath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRowsBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkDeleteRowsResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /v1/data/queries/{id}:
    post:
      tags: [Queries]
      summary: Execute a saved query
      operationId: executeQuery
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteQueryBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteQueryResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  parameters:
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: false
      description: |
        Tenant identifier. Required in multi-tenant mode (`TENANT_MODE=multi`) to route
        to the tenant's Postgres database. Also accepted as `X-Tenant-ID`.
      schema:
        type: string
        example: default
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    TableNamePath:
      name: tableName
      in: path
      required: true
      schema:
        type: string
    RowIdPath:
      name: rowId
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error message (plain text from http.Error)
      content:
        text/plain:
          schema:
            type: string
            example: table_name query parameter is required
  schemas:
    Value:
      type: object
      description: Cell value union (one field set per value).
      properties:
        stringValue:
          type: string
        numberValue:
          type: number
          format: double
        boolValue:
          type: boolean
        timestampValue:
          type: string
          format: date-time
        bytesValue:
          type: string
          format: byte
        jsonValue: {}
    Row:
      type: object
      properties:
        id:
          type: string
        cells:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Value"
...
    CreateRowBody:
      type: object
      properties:
        cells:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Value"
      description: Table id is taken from the path.
    CreateRowResponse:
      type: object
      properties:
        row:
          $ref: "#/components/schemas/Row"
    UpdateRowBody:
      type: object
      properties:
        cells:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Value"
      description: Table id and row id are taken from the path.
    UpdateRowResponse:
      type: object
      properties:
        row:
          $ref: "#/components/schemas/Row"
    DeleteRowResponse:
      type: object
    ListRowsResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/Row"
        nextPageToken:
          type: string
    SortOrder:
      type: object
      properties:
        attribute:
          type: string
        sortOrder:
          type: string
          enum: [ASC, DESC]
    QueryRowsBody:
      type: object
      properties:
        filter:
          type: object
          additionalProperties: true
        sort:
          type: array
          items:
            $ref: "#/components/schemas/SortOrder"
        columnIds:
          type: array
          items:
            type: string
        pageSize:
          type: integer
          format: int32
        pageToken:
          type: string
      description: Table id is taken from the path.
    QueryRowsResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/Row"
        nextPageToken:
          type: string
        count:
          type: integer
          format: int32
    BulkUpsertRowItem:
      type: object
      properties:
        rowId:
          type: string
        cells:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/Value"
    BulkUpsertRowsBody:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BulkUpsertRowItem"
      description: Table id is taken from the path.
    BulkUpsertRowsResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/Row"
    BulkDeleteRowsBody:
      type: object
      properties:
        rowIds:
          type: array
          items:
            type: string
      description: Table id is taken from the path.
    BulkDeleteRowsResponse:
      type: object
...
    ExecuteQueryBody:
      type: object
      properties:
        pageSize:
          type: integer
          format: int32
        pageToken:
          type: string
        filter:
          type: object
          additionalProperties: true
        params:
          type: object
          additionalProperties: true
          description: Replaces `{param}` placeholders in the saved query filter.
        columnIds:
          type: array
          items:
            type: string
          description: Optional subset of the query view projection (column_names). Omitted = return all columns defined on the query (empty column_names = SELECT *).
      description: Query name is taken from the path.
    ExecuteQueryResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/Row"
        nextPageToken:
          type: string
        count:
          type: integer
          format: int32
...